requirements.txt
pyxdg==0.28